Restructure project into backend and frontend
