+++ .env.example
@@ -1,5 +1,5 @@
 OPENAI_API_KEY=sk-xxxxxx
-UPLOAD_FOLDER=/app/data
+UPLOAD_FOLDER=/app/backend/data
 ALLOWED_EXTENSIONS=png,jpg,jpeg,webp
 MAX_FILE_SIZE_MB=8
 LOGIN_PASSWORD=API2025
