--- app/templates/result.html
@@ -1,29 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <meta charset="utf-8">
-    <title>Results</title>
-    <style>
-        body { font-family: Arial, sans-serif; margin:40px; }
-        pre { background:#f2f2f2; padding:10px; }
-        table { border-collapse:collapse; margin-bottom:20px; }
-        th, td { border:1px solid #ccc; padding:4px 8px; }
-    </style>
-</head>
-<body>
-    <a href="{{ url_for('upload') }}">Back</a>
-    {% for r in results %}
-    <h3>{{ r.filename }} - {{ r.job_id }}</h3>
-    <pre id="md{{ loop.index }}">{{ r.output }}</pre>
-    <button onclick="copy({{ loop.index }})">Copy Markdown</button>
-    <div>{{ r.output | safe }}</div>
-    <hr>
-    {% endfor %}
-<script>
-function copy(i){
-  let el = document.getElementById('md'+i);
-  navigator.clipboard.writeText(el.textContent);
-}
-</script>
-</body>
-</html>
