+++ frontend/index.html
@@ -3,13 +3,10 @@
 <head>
     <meta charset="utf-8">
     <title>Login</title>
-    <style>
-        body { font-family: Arial, sans-serif; display:flex; justify-content:center; align-items:center; height:100vh; background:#f2f2f2; }
-        form { background:#fff; padding:20px; border-radius:5px; box-shadow:0 2px 4px rgba(0,0,0,0.1); }
-    </style>
+    <link rel="stylesheet" href="{{ url_for('static', filename='style.css') }}">
 </head>
-<body>
-    <form method="post">
+<body class="login-page">
+    <form method="post" class="login">
         <h2>Login</h2>
         <input type="password" name="password" placeholder="Password" required/>
         <button type="submit">Enter</button>
