+++ frontend/main.js
@@ -1,0 +1,43 @@
+let dropArea = document.getElementById('drop-area');
+let fileElem = document.getElementById('fileElem');
+let gallery = document.getElementById('gallery');
+
+if (dropArea) {
+  ['dragenter', 'dragover', 'dragleave', 'drop'].forEach(eventName => {
+    dropArea.addEventListener(eventName, preventDefaults, false);
+  });
+  dropArea.addEventListener('drop', handleDrop, false);
+  fileElem.addEventListener('change', () => previewFiles(fileElem.files));
+}
+
+function preventDefaults (e) { e.preventDefault(); e.stopPropagation(); }
+
+function handleDrop(e) {
+  let dt = e.dataTransfer;
+  let files = dt.files;
+  fileElem.files = files;
+  previewFiles(files);
+}
+
+function previewFiles(files) {
+  gallery.innerHTML = '';
+  [...files].forEach(file => {
+    let img = document.createElement('img');
+    img.classList.add('thumb');
+    img.src = URL.createObjectURL(file);
+    gallery.appendChild(img);
+  });
+}
+
+function copy(i){
+  let el = document.getElementById('md'+i);
+  navigator.clipboard.writeText(el.textContent);
+}
+function download(i){
+  let el = document.getElementById('md'+i);
+  let blob = new Blob([el.textContent], {type: 'text/markdown'});
+  let a = document.createElement('a');
+  a.href = URL.createObjectURL(blob);
+  a.download = 'result_'+i+'.md';
+  a.click();
+}
