+++ frontend/upload.html
@@ -1,0 +1,23 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <meta charset="utf-8">
+    <title>Upload</title>
+    <link rel="stylesheet" href="{{ url_for('static', filename='style.css') }}">
+</head>
+<body>
+    <h2>Upload Images (Model: {{ model }})</h2>
+    <a href="{{ url_for('logout') }}">Logout</a>
+    <div id="drop-area">Drop images here</div>
+    <form id="form" method="post" enctype="multipart/form-data">
+        <input id="fileElem" type="file" name="files" accept="image/*" multiple style="display:none"/>
+        <button type="button" onclick="document.getElementById('fileElem').click()">Select Files</button>
+        <div id="gallery"></div>
+        <button type="submit">Upload</button>
+    </form>
+    {% with messages = get_flashed_messages() %}
+      {% if messages %}<div>{{ messages[0] }}</div>{% endif %}
+    {% endwith %}
+<script src="{{ url_for('static', filename='main.js') }}"></script>
+</body>
+</html>
