+++ requirements.txt
@@ -2,3 +2,4 @@ Flask
 python-dotenv
 Flask-Limiter
 openai
+markdown2
