+++ tests/test_basic.py
@@ -3,18 +3,23 @@ sys.path.append(str(pathlib.Path(__file__).resolve().parents[1]))
 import os
 import tempfile
 import pytest
-from app.main import app, init_db
+
+UPLOAD_DIR = tempfile.mkdtemp()
+os.environ.setdefault('UPLOAD_FOLDER', UPLOAD_DIR)
+from backend.app import app
+from backend.models import init_db
 
 @pytest.fixture
 def client():
-    db_fd, db_path = tempfile.mkstemp()
-    app.config['DATABASE'] = db_path
+    db_dir = UPLOAD_DIR
+    os.environ['UPLOAD_FOLDER'] = db_dir
     app.config['TESTING'] = True
     init_db()
     with app.test_client() as client:
         yield client
-    os.close(db_fd)
-    os.unlink(db_path)
+    for f in os.listdir(db_dir):
+        os.remove(os.path.join(db_dir, f))
+    os.rmdir(db_dir)
 
 
 def test_login(client):
